pyproject.toml
[project]
name = "educenso-analytics-api"
version = "0.1.0"
description = "Backend FastAPI para PI2 / EduCenso Analytics."
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0,<1.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.116.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "pydantic-settings>=2.10.0,<3.0.0",
    "python-jose[cryptography]>=3.5.0,<4.0.0",
    "sqlalchemy>=2.0.42,<3.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0,<9.0.0",
    "pytest-asyncio>=1.1.0,<2.0.0",
    "ruff>=0.12.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=80.0.0"]
build-backend = "setuptools.build_meta"
